+++ tests/test_end_to_end.py
@@ -45,6 +45,13 @@ class test_end_to_end(unittest.TestCase):
         images_equal = dice > 0.99
         self.assertTrue(images_equal, f"multilabel prediction not correct (dice: {dice:.6f})")
 
+    def test_prediction_multilabel_high_order(self):
+        img_ref = nib.load("tests/reference_files/example_seg.nii.gz").get_fdata()
+        img_new = nib.load("tests/unittest_prediction_high_order.nii.gz").get_fdata()
+        dice = dice_score_multilabel(img_ref, img_new)
+        images_equal = dice > 0.95
+        self.assertTrue(images_equal, f"multilabel prediction high order not correct (dice: {dice:.6f})")
+
     def test_prediction_multilabel_mr(self):
         img_ref = nib.load("tests/reference_files/example_seg_mr.nii.gz").get_fdata()
         img_new = nib.load("tests/unittest_prediction_mr.nii.gz").get_fdata()